feat: add hook on the run command.
