+++ cmd/cmd_run.go
@@ -39,6 +39,10 @@ func createRun() cli.Command {
 				Name:  "must-staple",
 				Usage: "Include the OCSP must staple TLS extension in the CSR and generated certificate. Only works if the CSR is generated by lego.",
 			},
+			cli.StringFlag{
+				Name:  "run-hook",
+				Usage: "Define a hook. The hook is executed when the certificates are effectively created.",
+			},
 		},
 	}
 }
@@ -86,7 +90,14 @@ func run(ctx *cli.Context) error {
 
 	certsStorage.SaveResource(cert)
 
-	return nil
+	meta := map[string]string{
+		renewEnvAccountEmail: account.Email,
+		renewEnvCertDomain:   cert.Domain,
+		renewEnvCertPath:     certsStorage.GetFileName(cert.Domain, ".crt"),
+		renewEnvCertKeyPath:  certsStorage.GetFileName(cert.Domain, ".key"),
+	}
+
+	return launchHook(ctx.String("run-hook"), meta)
 }
 
 func handleTOS(ctx *cli.Context, client *lego.Client) bool {
